Split reference preprocessing into separate script
